Optimize component structure
