+++ src/components/button/style/button.css
@@ -1,0 +1,119 @@
+@-webkit-keyframes rotation {
+    from {
+        transform: rotate(0deg);
+    }
+
+    to {
+        transform: rotate(360deg);
+    }
+}
+
+@keyframes rotation {
+    from {
+        transform: rotate(0deg);
+    }
+
+    to {
+        transform: rotate(360deg);
+    }
+}
+
+.nl-button {
+    --width: auto;
+    --height: 36px;
+
+    --border-color: #b9b9b9;
+    --background-color: transparent;
+
+    --font-size: 14px;
+    --font-color: inherit;
+    --font-family: "Consolas";
+    --font-weight: normal;
+
+    --icon-color: var(--font-color);
+    --icon-size: inherit;
+    --icon-weight: normal;
+
+    --hover-background-color: rgb(243, 243, 243);
+    --active-background-color: rgba(243, 243, 243, 0.36);
+
+    flex: none;
+    cursor: pointer;
+    user-select: none;
+    transition: all 0.24s ease-out;
+
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 6px;
+
+    box-sizing: border-box;
+    width: var(--width);
+    height: var(--height);
+    padding: 0 calc(var(--height) / 3 - 2px);
+
+    border: 1px solid var(--border-color);
+    background-color: var(--background-color);
+
+    color: var(--font-color);
+    font-size: var(--font-size);
+    font-family: var(--font-family);
+
+    & .iconfont {
+        color: var(--icon-color);
+        font-size: var(--icon-size);
+        font-weight: var(--icon-weight);
+    }
+
+    &:hover {
+        background-color: var(--hover-background-color);
+    }
+
+    &:active {
+        background-color: var(--active-background-color);
+    }
+
+    & > .nl-button__loading-icon {
+        -webkit-animation: 1.2s ease-in-out 0.16s infinite rotation;
+        animation: 1.2s ease-in-out 0.16s infinite rotation;
+    }
+
+    &.nl-button--square {
+        border-radius: 6px;
+    }
+
+    &.nl-button--round {
+        border-radius: var(--height);
+    }
+
+    &.nl-button--no-border {
+        border: none;
+        border-radius: 6px;
+    }
+
+    &.nl-button--round-no-border {
+        border: none;
+        border-radius: var(--height);
+    }
+
+    &.nl-button--no-shape {
+        --hover-background-color: transparent;
+        --active-background-color: transparent;
+
+        padding: 0;
+        margin: 0;
+        border: none;
+        border-radius: 0;
+    }
+
+    &.nl-button--disabled {
+        --font-color: #8e8e8e;
+        --background-color: var(--hover-background-color);
+        cursor: not-allowed;
+        opacity: 0.78;
+
+        &:active {
+            background-color: var(--hover-background-color);
+        }
+    }
+}
